requirements.txt
camelot-py==0.11.0
ghostscript
pdfplumber==0.9.0
loguru==0.7.0
langchain
transformers
sentencepiece==0.1.99
opencv-python
parse
fastbm25
pytest
tqdm
modelscope
vllm
peft
rouge_score
nltk
python-dotenv